// frontend/src/lib/api/model/documentChunkResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DocuMind.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface DocumentChunkResponse {
  index: number;
  pageNumber: number;
  endPageNumber: number;
  characterCount: number;
  tokenEstimate: number;
  overlapWithPrevious: number;
  text: string;
}